+++ main.py
@@ -961,7 +961,7 @@ class MainScreen(Screen):
             self.scenarioTitleLabel = Button(text=config.advDict[block]['title'], size_hint=(1,1), background_normal='', background_color=neutral, background_down='', background_color_down=accent2, font_name='Fantasque-Sans', font_size=config.basefont75)
             self.titleBarBox.add_widget(self.scenarioTitleLabel)
 
-            if config.scenario['block'] == 'Start' and config.advDict[block]['shown'] == 0:
+            if config.scenario['block'] == 'Start' and len(config.textLabelArray) == 1:
                 config.advDict[block]['shown'] == 99
                 showCurrentBlock(self)
                 saveconfig(self, config.curr_game_dir)